create the todo app
